refactor: extract tests for DHTProtocol from test_dht_node, re-write using asyncio
